rustmameui::games

Function add_favourite

pub fn add_favourite(
    config: &Config,
    favourites: &mut Vec<Game>,
    game: &Game,
) -> Vec<Game>

Adds a Game to the vector of favourites and saves the updated list to the favourites.json file.

Arguments

config - A reference to the application Config, containing the project configuration directory path.

favourites - A mutable reference to the current vector of favourite Games. The game will be added to this vector.

game - A reference to the Game to add to favourites.

Returns

A clone of the updated favourites vector after adding the game and attempting to save. Errors during file creation or writing are printed to eprintln!, and the function still returns the (potentially unsaved) updated vector. Serialization errors are also printed.