rustmameui::games

Function save

pub fn save(config: &Config, games: &Vec<Game>) -> Result<()>

Saves a vector of Game structs to a JSON file in the application’s configuration directory.

The file is named games.json within the directory specified by config.project_config_dir. The data is serialized with pretty printing.

Arguments

config - A reference to the application Config, containing the project configuration directory path.

games - A reference to a vector of Game structs to save.

Returns

A std::io::Result<()> indicating success or failure. Returns an error if the file cannot be created, the data cannot be serialized, or the data cannot be written to the file.